misc: add dynamic writeback interface
Before this patch, the deamon for flushing
dirty pages kicks in every 600 ms. This patch
changes the behaviour a little bit. While suspended
it flushes more pages (every 500 ms) and while the
the device is awake and in use it flushes only every
4 s to gain a little performance boost.

This is fully controllable;

echo 1 > /sys/class/misc/wirteback/writeback_enabled
(1 = On (default), 0 = off (and behaviour before this
patch).

Signed-off-by: Blechd0se <alex.christ@hotmail.de>

Conflicts:
	arch/arm/configs/mapphone_defconfig
	drivers/misc/Kconfig
	drivers/misc/Makefile
	mm/page-writeback.c